+++ packages/core/client/src/block-provider/DetailsBlockProvider.tsx
@@ -2,13 +2,11 @@ import { createForm } from '@formily/core';
 import { useField, useFieldSchema } from '@formily/react';
 import { Spin } from 'antd';
 import React, { createContext, useContext, useEffect, useMemo, useRef } from 'react';
-import { useTranslation } from 'react-i18next';
 import { withDynamicSchemaProps } from '../application/hoc/withDynamicSchemaProps';
 import { useCollectionManager_deprecated } from '../collection-manager';
-import { useCollection, useCollectionRecordData } from '../data-source';
+import { useCollectionRecordData } from '../data-source';
 import { useCollectionParentRecord } from '../data-source/collection-record/CollectionRecordProvider';
 import { useDetailsWithPaginationBlockParams } from '../modules/blocks/data-blocks/details-multi/hooks/useDetailsWithPaginationBlockParams';
-import { DeclareVariable } from '../modules/variable/DeclareVariable';
 import { RecordProvider } from '../record-provider';
 import { useDesignable } from '../schema-component';
 import { BlockProvider, useBlockRequestContext } from './BlockProvider';
@@ -22,8 +20,6 @@ DetailsBlockContext.displayName = 'DetailsBlockContext';
 
 const InternalDetailsBlockProvider = (props) => {
   const { action, readPretty } = props;
-  const { t } = useTranslation();
-  const collection = useCollection();
   const field = useField<any>();
   const form = useMemo(
     () =>
@@ -56,14 +52,7 @@ const InternalDetailsBlockProvider = (props) => {
     <DetailsBlockContext.Provider value={detailsBLockValue}>
       <div ref={formBlockRef}>
         <RecordProvider isNew={false} record={currentRecord} parent={parentRecord?.data}>
-          <DeclareVariable
-            name="$nPopupRecord"
-            title={t('Current popup record')}
-            value={currentRecord}
-            collection={collection}
-          >
-            {props.children}
-          </DeclareVariable>
+          {props.children}
         </RecordProvider>
       </div>
     </DetailsBlockContext.Provider>
