+++ packages/core/client/src/schema-component/antd/grid-card/GridCard.Item.tsx
@@ -3,11 +3,8 @@ import { ObjectField } from '@formily/core';
 import { useField } from '@formily/react';
 import { Card } from 'antd';
 import React from 'react';
-import { useTranslation } from 'react-i18next';
 import { withDynamicSchemaProps } from '../../../application/hoc/withDynamicSchemaProps';
 import { useCollectionParentRecordData } from '../../../data-source/collection-record/CollectionRecordProvider';
-import { useCollection } from '../../../data-source/collection/CollectionProvider';
-import { DeclareVariable } from '../../../modules/variable/DeclareVariable';
 import { RecordProvider } from '../../../record-provider';
 
 const itemCss = css`
@@ -32,22 +29,13 @@ const gridCardCss = css`
 
 export const GridCardItem = withDynamicSchemaProps(
   (props) => {
-    const { t } = useTranslation();
-    const collection = useCollection();
     const field = useField<ObjectField>();
     const parentRecordData = useCollectionParentRecordData();
     return (
       <Card role="button" aria-label="grid-card-item" className={gridCardCss}>
         <div className={itemCss}>
           <RecordProvider record={field.value} parent={parentRecordData}>
-            <DeclareVariable
-              name="$nPopupRecord"
-              title={t('Current popup record')}
-              value={field.value}
-              collection={collection}
-            >
-              {props.children}
-            </DeclareVariable>
+            {props.children}
           </RecordProvider>
         </div>
       </Card>
